fix: auto-convert INTEGER to BIGINT on PostgreSQL at execute level
Migration files use INTEGER PRIMARY KEY (which gives rowid alias on
SQLite for auto-increment). On PostgreSQL, INTEGER cannot hold 64-bit
snowflake IDs. Added apply_pg_type_fixes() in the database execute
path that auto-converts INTEGER->BIGINT and BOOLEAN DEFAULT 0/1 -> FALSE/TRUE
for all DDL on PostgreSQL.

Also fixed automod/schema.py: cast BIGINT column to text before applying
~ regex operator (PG16 rejects 'operator does not exist: bigint ~ unknown').

bump version to a.1.0-73

+++ src/core/migrations/migrations/008_harden_admin_security.py
@@ -79,7 +79,7 @@ def up(db):
         db.execute("""
             CREATE TABLE IF NOT EXISTS admin_notes (
                 id INTEGER PRIMARY KEY,
-                ticket_id BIGINT NOT NULL,
+                ticket_id INTEGER NOT NULL,
                 admin_id INTEGER NOT NULL,
                 content TEXT NOT NULL,
                 created_at INTEGER NOT NULL,
@@ -91,7 +91,7 @@ def up(db):
         db.execute("""
             CREATE TABLE IF NOT EXISTS admin_notes (
                 id BIGINT PRIMARY KEY,
-                ticket_id BIGINT NOT NULL,
+                ticket_id INTEGER NOT NULL,
                 admin_id BIGINT NOT NULL,
                 content TEXT NOT NULL,
                 created_at BIGINT NOT NULL,
